refactor: InnerClass 분리
